Refactor Teaser class and extract WidgetTrait for better code organization
